+++ serwis_panel/src/main/java/org/pl/serwis_panel/repositories/UserRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     User getUserByLogin(String login);
-    User getUserById(int id);
+    User getUserById(Long id);
 }
