+++ serwis_panel/src/main/java/org/pl/serwis_panel/services/AdminPanelService.java
@@ -17,7 +17,7 @@ public class AdminPanelService {
         return userRepository.getUserByLogin(login);
     }
 
-    public User getUserById(int id) {
+    public User getUserById(Long id) {
         return userRepository.getUserById(id);
     }
